refactor: 외부 기술에 의존적인 도메인 infra 패키지로 분리
